+++ CHANGELOG.md
@@ -9,8 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Tenant id is taken from JWT token (default is 'd4l')
-
 ### Changed
 
 ### Deprecated
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v0.10.1] - 2020-12-16
+
+### Added
+
+- Tenant id is taken from JWT token (default is 'd4l')
 
 ## [v0.10.0] - 2020-11-24
 
@@ -126,7 +130,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v0.10.1...HEAD
+[v0.10.1]: https://github.com/gesundheitscloud/go-svc/compare/v0.10.0...v0.10.1
 [v0.10.0]: https://github.com/gesundheitscloud/go-svc/compare/v0.9.2...v0.10.0
 [v0.9.2]: https://github.com/gesundheitscloud/go-svc/compare/v0.9.1...v0.9.2
 [v0.9.1]: https://github.com/gesundheitscloud/go-svc/compare/v0.9.0...v0.9.1